feat: implementa context
