+++ src/model/user.model.ts
@@ -1,0 +1,4 @@
+export interface UserModel {
+  name: string;
+  email: string;
+}
